Initial commit - At least of the refactor
This contains a refactor of the image sorting and detail population. The
file sorter now takes a copy of the original files and places them in a
staging area. they are then renamed an put on the same top level.
Information surrounding the image properties is also populated as part
of this process.